unsigned characteristics

+++ tests/test-fgemm-long.C
@@ -280,9 +280,9 @@ bool launch_MM_dispatch(const Field &F,
 	return ok ;
 }
 template <class Field>
-bool run_with_field (int q, unsigned long b, size_t n, int nbw, size_t iters, bool par ){
+bool run_with_field (unsigned long q, unsigned long b, size_t n, int nbw, size_t iters, bool par ){
 	bool ok = true ;
-	int p=q;
+	unsigned long p=q;
 	int nbit=(int)iters;
 	while (ok &&  nbit){
 		typedef typename  Field::Element Element ;
